move StagedRegion
