+++ hail/src/main/scala/is/hail/annotations/Region.scala
@@ -1,8 +1,7 @@
 package is.hail.annotations
 
 import is.hail.asm4s
-import is.hail.asm4s.{Code, CodeBuilder, MethodBuilder, Settable, Value, coerce}
-import is.hail.expr.ir.EmitMethodBuilder
+import is.hail.asm4s.{Code, coerce}
 import is.hail.types.physical._
 import is.hail.utils._
 
@@ -491,80 +490,4 @@ object RegionUtils {
 
   def logRegionStats(header: String, region: Code[Region]): Code[Unit] =
     Code.invokeScalaObject2[String, Region, Unit](RegionUtils.getClass, "logRegionStats", header, region)
-}
-
-abstract class StagedRegion {
-  def code: Value[Region]
-
-  def createChildRegion(mb: EmitMethodBuilder[_], allowSubregions: Boolean = false): StagedOwnedRegion
-}
-
-trait StagedOwnedRegion extends StagedRegion {
-  def allocateRegion(size: Int): Code[Unit]
-
-  def free(): Code[Unit]
-
-  def clear(): Code[Unit]
-
-  def giveToParent(): Code[Unit]
-
-  def shareWithParent(): Code[Unit]
-
-  def copyToParent(mb: EmitMethodBuilder[_], value: PCode): PCode
-}
-
-object StagedRegion {
-  def apply(r: Value[Region], allowSubregions: Boolean = false): StagedRegion =
-    if (allowSubregions) new RealStagedRegion(r) else new DummyStagedRegion(r)
-
-  class RealStagedRegion(r: Value[Region]) extends StagedRegion {
-    def code: Value[Region] = r
-
-    def createChildRegion(mb: EmitMethodBuilder[_], allowSubregions: Boolean = false): StagedOwnedRegion = {
-      if (allowSubregions) {
-        val newR = mb.newLocal[Region]("staged_region_child")
-        new RealStagedOwnedRegion(newR, r)
-      } else {
-        new DummyStagedRegion(r)
-      }
-    }
-  }
-
-  class RealStagedOwnedRegion(r: Settable[Region], parent: Value[Region]) extends RealStagedRegion(r) with StagedOwnedRegion {
-    def allocateRegion(size: Int): Code[Unit] = r := Region.stagedCreate(size)
-
-    def free(): Code[Unit] = Code(r.invalidate(), r := Code._null)
-
-    def clear(): Code[Unit] = (r: Value[Region]).clear()
-
-    def giveToParent(): Code[Unit] = parent.invoke("takeOwnershipOf", r)
-
-    def shareWithParent(): Code[Unit] = parent.invoke("addReferenceTo", r)
-
-    def copyToParent(mb: EmitMethodBuilder[_], value: PCode): PCode =
-      value.copyToRegion(mb, parent)
-  }
-
-  class DummyStagedRegion(r: Value[Region]) extends StagedOwnedRegion {
-    def code: Value[Region] = r
-
-    def createChildRegion(mb: EmitMethodBuilder[_], allowSubregions: Boolean = false): StagedOwnedRegion =
-      this
-
-    def allocateRegion(size: Int): Code[Unit] = Code._empty
-
-    def free(): Code[Unit] = Code._empty
-
-    def clear(): Code[Unit] = Code._empty
-
-    def giveToParent(): Code[Unit] = Code._empty
-
-    def shareWithParent(): Code[Unit] = Code._empty
-
-    def copyToParent(mb: EmitMethodBuilder[_], value: PCode): PCode = value
-  }
-}
-
-abstract class StagedRegionPool {
-  def getRegion(): Code[Region]
 }
